Delete migration task UpdateChargingStationTemplate, use the scheduler one named CheckChargingStationTemplateTask
