+++ src/scheduler/tasks/CheckChargingStationTemplateTask.ts
@@ -6,6 +6,7 @@ import Constants from '../../utils/Constants';
 import { LockEntity } from '../../types/Locking';
 import LockingManager from '../../locking/LockingManager';
 import Logging from '../../utils/Logging';
+import { OCPPPhase } from '../../types/ocpp/OCPPServer';
 import OCPPUtils from '../../server/ocpp/utils/OCPPUtils';
 import SchedulerTask from '../SchedulerTask';
 import { ServerAction } from '../../types/Server';
@@ -67,6 +68,17 @@ export default class CheckChargingStationTemplateTask extends SchedulerTask {
         for (const connector of chargingStation.connectors) {
           if (!Utils.objectHasProperty(connector, 'amperageLimit')) {
             connector.amperageLimit = connector.amperage;
+            chargingStationUpdated = true;
+          }
+          if (!Utils.objectHasProperty(connector, 'phaseAssignmentToGrid')) {
+            const numberOfPhases = Utils.getNumberOfConnectedPhases(chargingStation, null, connector.connectorId);
+            if (numberOfPhases === 1) {
+              connector.phaseAssignmentToGrid = { csPhaseL1: OCPPPhase.L1, csPhaseL2: null, csPhaseL3: null } ;
+            } else if (numberOfPhases === 3) {
+              connector.phaseAssignmentToGrid = { csPhaseL1: OCPPPhase.L1, csPhaseL2: OCPPPhase.L2, csPhaseL3: OCPPPhase.L3 } ;
+            } else {
+              connector.phaseAssignmentToGrid = null;
+            }
             chargingStationUpdated = true;
           }
         }
